hassan

+++ libraries/libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yumi <yumi@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: aelkheta <aelkheta@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 12:16:43 by aelkheta          #+#    #+#             */
-/*   Updated: 2024/05/27 09:00:26 by yumi             ###   ########.fr       */
+/*   Updated: 2024/08/09 16:28:14 by aelkheta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ char				*get_the_rest(char *buffer);
 char				*check_next_line(char *buffer);
 char				*reach_the_line(int fd, char *buffer);
 
-long int			ft_atoi(const char *str);
+long long			ft_atoi(const char *str);
 void				ft_bzero(void *s, size_t n);
 void				*ft_calloc(size_t n, size_t size);
 int					ft_isalnum(int c);
